Adds var guards to find fns

+++ shared/required/home/.bash_functions
@@ -315,6 +315,7 @@ function refresh_pwsh {
 
 function find_items_rename_experimental_helper {
   local file_pattern="$1";
+  [[ -z "$file_pattern" ]] && { echo "Must specifiy a file pattern!"; return 1; }
   local by="$2";
   local preview=$3;
   [[ -z "$preview" ]] && { echo "Must specifiy the preview flag!"; return 1; }
@@ -346,6 +347,7 @@ function find_items_rename_experimental_helper {
 
 function find_items_rename_preview_experimental {
   local file_pattern="$1";
+  [[ -z "$file_pattern" ]] && { echo "Must specifiy a file pattern!"; return 1; }
   local by="$2";
   local preview=true
   find_items_rename_experimental_helper "$file_pattern" "$by" $preview;
@@ -353,6 +355,7 @@ function find_items_rename_preview_experimental {
 
 function find_items_rename_experimental {
   local file_pattern="$1";
+  [[ -z "$file_pattern" ]] && { echo "Must specifiy a file pattern!"; return 1; }
   local by="$2";
   local preview=false
   find_items_rename_experimental_helper "$file_pattern" "$by" $preview;
@@ -360,6 +363,7 @@ function find_items_rename_experimental {
 
 function find_items {
   local file_pattern="$1";
+  [[ -z "$file_pattern" ]] && { echo "Must specifiy a file pattern!"; return 1; }
   find . -maxdepth 9 -regextype egrep -iregex "$file_pattern" -not -path '*/__pycache__/*' -not -path '*/bin/*' -not -path '*/obj/*' -not -path '*/.git/*' -not -path '*/.svn/*' -not -path '*/node_modules/*' -not -path '*/.ionide/*'
 }
 
@@ -370,6 +374,7 @@ function find_items_fuzz {
 
 function find_files_rename_helper {
   local file_pattern="$1";
+  [[ -z "$file_pattern" ]] && { echo "Must specifiy a file pattern!"; return 1; }
   local by="$2";
   local preview=$3
   [[ -z "$preview" ]] && { echo "Must specifiy the preview flag!"; return 1; }
@@ -390,6 +395,7 @@ function find_files_rename_helper {
 
 function find_files_delete_preview {
   local file_pattern="$1";
+  [[ -z "$file_pattern" ]] && { echo "Must specifiy a file pattern!"; return 1; }
   local with_content="$2";
   [[ -z "$with_content" ]] && {
     find . -maxdepth 9 -regextype egrep -iregex "$file_pattern" -type f -not -path '*/__pycache__/*' -not -path '*/bin/*' -not -path '*/obj/*' -not -path '*/.git/*' -not -path '*/.svn/*' -not -path '*/node_modules/*' -not -path '*/.ionide/*' -exec echo rm "{}" \;
@@ -400,6 +406,7 @@ function find_files_delete_preview {
 
 function find_files_delete {
   local file_pattern="$1";
+  [[ -z "$file_pattern" ]] && { echo "Must specifiy a file pattern!"; return 1; }
   local with_content="$2";
   [[ -z "$with_content" ]] && {
     find . -maxdepth 9 -regextype egrep -iregex "$file_pattern" -type f -not -path '*/__pycache__/*' -not -path '*/bin/*' -not -path '*/obj/*' -not -path '*/.git/*' -not -path '*/.svn/*' -not -path '*/node_modules/*' -not -path '*/.ionide/*' -exec rm "{}" \;
@@ -410,6 +417,7 @@ function find_files_delete {
 
 function find_files_rename_preview {
   local file_pattern="$1";
+  [[ -z "$file_pattern" ]] && { echo "Must specifiy a file pattern!"; return 1; }
   local by="$2";
   local preview=true;
   find_files_rename_helper "$file_pattern" "$by" "$preview";
@@ -417,6 +425,7 @@ function find_files_rename_preview {
 
 function find_files_rename {
   local file_pattern="$1";
+  [[ -z "$file_pattern" ]] && { echo "Must specifiy a file pattern!"; return 1; }
   local by="$2";
   local preview=false;
   find_files_rename_helper "$file_pattern" "$by" "$preview";
@@ -424,6 +433,7 @@ function find_files_rename {
 
 function find_files {
   local file_pattern="$1";
+  [[ -z "$file_pattern" ]] && { echo "Must specifiy a file pattern!"; return 1; }
   find . -maxdepth 9 -regextype egrep -iregex "$file_pattern" -type f -not -path '*/__pycache__/*' -not -path '*/bin/*' -not -path '*/obj/*' -not -path '*/.git/*' -not -path '*/.svn/*' -not -path '*/node_modules/*' -not -path '*/.ionide/*';
 }
 
@@ -434,6 +444,7 @@ function find_files_fuzz {
 
 function find_in_files {
   local grep_pattern="$1";
+  [[ -z "$grep_pattern" ]] && { echo "Must specifiy a grep pattern!"; return 1; }
   local file_pattern="$2";
   [[ -z "$file_pattern" ]] && { file_pattern=".*"; }
   find . -maxdepth 9 -regextype egrep -iregex "$file_pattern" -type f -not -path '*/__pycache__/*' -not -path '*/bin/*' -not -path '*/obj/*' -not -path '*/.git/*' -not -path '*/.svn/*' -not -path '*/node_modules/*' -not -path '*/.ionide/*' -exec egrep --color -in "$grep_pattern" "{}" +;
@@ -448,6 +459,7 @@ function find_in_files_fuzz {
 
 function find_in_files_replace {
   local by="$1"
+  [[ -z "$by" ]] && { echo "Must specifiy a by substitution!"; return 1; }
   local file_pattern="$2";
   [[ -z "$file_pattern" ]] && { file_pattern=".*"; }
   find . -maxdepth 9 -regextype egrep -iregex "$file_pattern" -type f -not -path '*/__pycache__/*' -not -path '*/bin/*' -not -path '*/obj/*' -not -path '*/.git/*' -not -path '*/.svn/*' -not -path '*/node_modules/*' -not -path '*/.ionide/*' -exec sed -E -i'' "$by" "{}" \;
@@ -479,6 +491,7 @@ function git_log_follow {
 
 function show_cmds_like {
   local pattern="$1";
+  [[ -z "$pattern" ]] && { echo "Must specifiy a command pattern!"; return 1; }
   local search_res=$(search_env_for "$pattern");
   local alias=$(echo "$search_res" | egrep -i "\s*alias");
   [[ -z "$alias" ]] || { echo "$alias"; }
